requirements.txt
langchain==0.2.1
langchain-core==0.2.0
langchain-google-vertexai
google-cloud-aiplatform
google-auth==2.29.0
pydantic==1.10.9
python-dotenv==1.0.1
streamlit==1.35.0
